requirements.txt
annotated-types==0.7.0
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
langcodes==3.5.1
lxml==6.0.2
MarkupSafe==3.0.3
murmurhash==1.0.15
nltk==3.8.1
numpy==1.26.4
packaging==26.0
pandas==2.1.4
pdfminer.six==20231228
preshed==3.0.12
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2025.2
regex==2026.1.15
requests==2.32.5
scikit-learn==1.3.2
scipy==1.17.0
six==1.17.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
textblob==0.17.1
thinc==8.2.5
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.9.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wasabi==1.1.3
weasel==0.3.4
Werkzeug==3.1.5
